development-process/stable-maintenance/regression-analysis/requirements.txt
# SPDX-FileCopyrightText: 2019 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
#
# SPDX-License-Identifier: GPL-2.0-only

gitpython
pandas
plotly
tabulate
